feat(api): move GET api/admin/organizations/id/children to right context
