+++ api/tests/organizational-entities/acceptance/application/organization/organization.admin.route.test.js
@@ -1,4 +1,5 @@
 import iconv from 'iconv-lite';
+import lodash from 'lodash';
 
 import { PIX_ADMIN } from '../../../../../src/authorization/domain/constants.js';
 import { ORGANIZATION_FEATURE } from '../../../../../src/shared/domain/constants.js';
@@ -13,6 +14,7 @@ import {
 } from '../../../../test-helper.js';
 
 const { ROLES } = PIX_ADMIN;
+const { map: _map } = lodash;
 
 describe('Acceptance | Organizational Entities | Application | Route | Admin | Organization', function () {
   let admin;
@@ -24,6 +26,85 @@ describe('Acceptance | Organizational Entities | Application | Route | Admin | O
     await databaseBuilder.commit();
 
     server = await createServer();
+  });
+
+  describe('GET /api/admin/organizations/{organizationId}/children', function () {
+    context('error cases', function () {
+      context('when organization does not exist', function () {
+        it('returns a 404 HTTP status code with an error message', async function () {
+          // given
+          const userId = databaseBuilder.factory.buildUser.withRole().id;
+
+          await databaseBuilder.commit();
+
+          const request = {
+            method: 'GET',
+            url: '/api/admin/organizations/986532/children',
+            headers: generateAuthenticatedUserRequestHeaders({ userId }),
+          };
+
+          // when
+          const response = await server.inject(request);
+
+          // then
+          expect(response.statusCode).to.equal(404);
+          expect(response.result.errors[0].detail).to.equal('Organization with ID (986532) not found');
+        });
+      });
+
+      context('when the user does not have access to the resource', function () {
+        it('returns a 403 HTTP status code with an error message', async function () {
+          // given
+          const userId = databaseBuilder.factory.buildUser().id;
+          const organizationId = databaseBuilder.factory.buildOrganization().id;
+
+          await databaseBuilder.commit();
+
+          const request = {
+            method: 'GET',
+            url: `/api/admin/organizations/${organizationId}/children`,
+            headers: generateAuthenticatedUserRequestHeaders({ userId }),
+          };
+
+          // when
+          const response = await server.inject(request);
+
+          // then
+          expect(response.statusCode).to.equal(403);
+        });
+      });
+    });
+
+    context('success cases', function () {
+      // eslint-disable-next-line mocha/no-setup-in-describe
+      Object.keys(ROLES).forEach((role) => {
+        context(`when user has role ${role}`, function () {
+          it('returns child organizations list with a 200 HTTP status code', async function () {
+            // given
+            const userId = databaseBuilder.factory.buildUser.withRole({ role }).id;
+            const parentOrganizationId = databaseBuilder.factory.buildOrganization().id;
+
+            const firstChildId = databaseBuilder.factory.buildOrganization({ parentOrganizationId }).id + '';
+            const secondChildId = databaseBuilder.factory.buildOrganization({ parentOrganizationId }).id + '';
+
+            await databaseBuilder.commit();
+
+            const request = {
+              method: 'GET',
+              url: `/api/admin/organizations/${parentOrganizationId}/children`,
+              headers: generateAuthenticatedUserRequestHeaders({ userId }),
+            };
+            // when
+            const response = await server.inject(request);
+
+            // then
+            expect(response.statusCode).to.equal(200);
+            expect(response.result.data).to.have.lengthOf(2);
+            expect(_map(response.result.data, 'id')).to.have.members([firstChildId, secondChildId]);
+          });
+        });
+      });
+    });
   });
 
   describe('GET /api/admin/organizations', function () {
